requirements.txt
attrs==21.4.0
bcrypt==3.2.0
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
dnspython==2.2.1
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
git-filter-repo==2.34.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
is-disposable-email==1.0.0
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pyparsing==3.0.8
pytest==7.1.1
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.35
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
WTForms==3.0.1
wtforms-validators==1.0.0
zipp==3.8.0
